Requesting from server each 5 seconds; Basic Settings functionality are working;
Also, I did some code refactor again. I changed few things regarding the objects creation, (i.e. blueRoute and greenRoute). Now, we are creating just once.  Furthermore, we have the Map as an global object, so we don't need to use it as methods arguments as well as we don't need to keep creating it in other methods.