Added routes 1A and 1B. (A review is needed).
Now, the routes are more general, each one has its own color, center point and zoom.
To change the route colors, go to res/values/color.xml.

App.java included, so if we need to use the main context, we don't need to pass it as a parameter anymore. Instead, use App.getContext().