Changed Drawer color and poor version of the closest Bus Stops.
I changed the color of the drawer navigation bar to black again.
MyLocation now shows all the bus stops that we have in the routes. If we click in the bus stop, it's going to show the route that pass there. We need to change the way that it is shown as well as create a mode that we can see more than one route for each bus stop (when applied).